refactor: centralize orchestration
